+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added [Get-Histogram][] `-SkipBlank` option.
 - Added [Measure-Summary][] Count-NA feature.
+- Added [Map-Object][] `-SortAs <type>` option.
 
 ## [0.7.0] - 2025-04-12